admin controll

+++ routes/tasks.py
@@ -29,21 +29,33 @@ def create_task():
 @task_bp.route('/', methods=['GET'])
 @jwt_required()
 def get_tasks():
-    user_id = get_jwt_identity()['user_id']
-    tasks = Task.query.filter_by(user_id=user_id).all()
+    current_user = get_jwt_identity()
+    user_id = current_user['user_id']
+    is_admin = current_user['is_admin']
+    if is_admin:
+        tasks = Task.query.all()
+    else:
+        tasks = Task.query.filter_by(user_id=user_id).all()
+
     return jsonify([{
         'id': task.id,
         'title': task.title,
         'description': task.description,
-        'reminder_time': task.reminder_time.isoformat() if task.reminder_time else None # noqa
+        'reminder_time': task.reminder_time.isoformat() if task.reminder_time else None  # noqa
     } for task in tasks]), 200
 
 
 @task_bp.route('/<int:task_id>', methods=['GET'])
 @jwt_required()
 def get_task(task_id):
-    user_id = get_jwt_identity()['user_id']
-    task = Task.query.filter_by(id=task_id, user_id=user_id).first()
+    current_user = get_jwt_identity()
+    user_id = current_user['user_id']
+    is_admin = current_user['is_admin']
+
+    if is_admin:
+        task = Task.query.filter_by(id=task_id).first()
+    else:
+        task = Task.query.filter_by(id=task_id, user_id=user_id).first()
 
     if task is None:
         return jsonify({"msg": "Task not found"}), 404
@@ -52,18 +64,25 @@ def get_task(task_id):
         'id': task.id,
         'title': task.title,
         'description': task.description,
-        'reminder_time': task.reminder_time.isoformat() if task.reminder_time else None # noqa
+        'reminder_time': task.reminder_time.isoformat() if task.reminder_time else None  # noqa
     }), 200
 
 
 @task_bp.route('/<int:task_id>', methods=['PUT'])
 @jwt_required()
 def update_task(task_id):
-    user_id = get_jwt_identity()['user_id']
-    task = Task.query.filter_by(id=task_id, user_id=user_id).first()
+    current_user = get_jwt_identity()
+    user_id = current_user['user_id']
+    is_admin = current_user['is_admin']
+
+    task = Task.query.filter_by(id=task_id).first()
 
     if task is None:
         return jsonify({"msg": "Task not found"}), 404
+
+    if not is_admin and task.user_id != user_id:
+        return jsonify({"msg":
+                        "You do not have permission to update this task"}), 403
 
     title = request.json.get('title', task.title)
     description = request.json.get('description', task.description)
@@ -83,11 +102,18 @@ def update_task(task_id):
 @task_bp.route('/<int:task_id>', methods=['DELETE'])
 @jwt_required()
 def delete_task(task_id):
-    user_id = get_jwt_identity()['user_id']
-    task = Task.query.filter_by(id=task_id, user_id=user_id).first()
+    current_user = get_jwt_identity()
+    user_id = current_user['user_id']
+    is_admin = current_user['is_admin']
+
+    task = Task.query.filter_by(id=task_id).first()
 
     if task is None:
         return jsonify({"msg": "Task not found"}), 404
+
+    if not is_admin and task.user_id != user_id:
+        return jsonify({"msg":
+                        "You do not have permission to delete this task"}), 403
 
     db.session.delete(task)
     db.session.commit()
